+++ src/zenml/zen_server/routers/server_endpoints.py
@@ -17,7 +17,7 @@ from fastapi import APIRouter
 
 import zenml
 from zenml.constants import API, INFO, VERSION_1
-from zenml.models.server_models import ServerModel
+from zenml.models import ServerModel
 from zenml.zen_server.exceptions import error_response
 from zenml.zen_server.utils import handle_exceptions, zen_store
